% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acorn.R
\name{get_acorn}
\alias{get_acorn}
\title{Title}
\usage{
get_acorn(dest)
}
\arguments{
\item{dest}{Directory to extract into}
}
\value{
Nothing
}
\description{
Title
}
\examples{
\dontrun{
get_acorn("path/to/directory")
}
}
