% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_acorn.R
\name{merge_acorn}
\alias{merge_acorn}
\title{Read and merge the CSV files for every Station}
\usage{
merge_acorn(dir)
}
\arguments{
\item{dir}{Directory containing the desired CSV files}
}
\value{
A tibble
}
\description{
Read and merge the CSV files for every Station
}
\examples{
\dontrun{
# merge all stations:
merge_acorn("path/to/directory")
}
}
