% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_station.R
\name{read_station}
\alias{read_station}
\title{Read a station's CSV file}
\usage{
read_station(file)
}
\arguments{
\item{file}{Path and filename of CSV file to read (as a string)}
}
\value{
A clean tibble.
}
\description{
This function imports and clean a single ACORN station CSV file.
}
\examples{
\dontrun{
# read one station:
read_station("path/to/file.csv")
}
}
